+++ contracts/policies/CombinedPoliciesPolicy.sol
@@ -37,9 +37,9 @@ contract CombinedPoliciesPolicy is FirewallPolicyBase {
      * @param data The data of the transaction.
      * @param value The value of the transaction.
      */
-    function preExecution(address consumer, address sender, bytes calldata data, uint value) external isAuthorized(consumer) {
+    function preExecution(address consumer, address sender, bytes calldata data, uint256 value) external isAuthorized(consumer) {
         bool[] memory currentResult = new bool[](policies.length);
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             IFirewallPolicy policy = IFirewallPolicy(policies[i]);
             try policy.preExecution(consumer, sender, data, value) {
                 currentResult[i] = true;
@@ -59,10 +59,10 @@ contract CombinedPoliciesPolicy is FirewallPolicyBase {
      * @param data The data of the transaction.
      * @param value The value of the transaction.
      */
-    function postExecution(address consumer, address sender, bytes calldata data, uint value) external isAuthorized(consumer) {
+    function postExecution(address consumer, address sender, bytes calldata data, uint256 value) external isAuthorized(consumer) {
         bool[] memory currentResult = currentResults[currentResults.length - 1];
         currentResults.pop();
-        for (uint i = 0; i < policies.length; i++) {
+        for (uint256 i = 0; i < policies.length; i++) {
             IFirewallPolicy policy = IFirewallPolicy(policies[i]);
             try policy.postExecution(consumer, sender, data, value) {
                 // Do nothing
@@ -82,12 +82,12 @@ contract CombinedPoliciesPolicy is FirewallPolicyBase {
      */
     function setAllowedCombinations(address[] calldata _policies, bool[][] calldata _allowedCombinations) external onlyRole(POLICY_ADMIN_ROLE) {
         // Reset all combinations to false
-        for (uint i = 0; i < allowedCombinationHashes.length; i++) {
+        for (uint256 i = 0; i < allowedCombinationHashes.length; i++) {
             isAllowedCombination[allowedCombinationHashes[i]] = false;
         }
         allowedCombinationHashes = new bytes32[](_allowedCombinations.length);
         // Set all new combinations to true
-        for (uint i = 0; i < _allowedCombinations.length; i++) {
+        for (uint256 i = 0; i < _allowedCombinations.length; i++) {
             require(_policies.length == _allowedCombinations[i].length, "CombinedPoliciesPolicy: Invalid combination length");
             isAllowedCombination[keccak256(abi.encodePacked(_allowedCombinations[i]))] = true;
             allowedCombinationHashes[i] = (keccak256(abi.encodePacked(_allowedCombinations[i])));
