+++ contracts/samples/SampleConsumerUpgradeable.sol
@@ -8,8 +8,8 @@ import {Address} from "@openzeppelin/contracts/utils/Address.sol";
 import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 contract SampleConsumerUpgradeable is OwnableUpgradeable {
-    mapping (address => uint) public deposits;
-    mapping (address => mapping (address => uint)) public tokenDeposits;
+    mapping (address => uint256) public deposits;
+    mapping (address => mapping (address => uint256)) public tokenDeposits;
 
     function initialize() external initializer {
         __Ownable_init();
@@ -19,17 +19,17 @@ contract SampleConsumerUpgradeable is OwnableUpgradeable {
         deposits[msg.sender] += msg.value;
     }
 
-    function withdraw(uint amount) external {
+    function withdraw(uint256 amount) external {
         deposits[msg.sender] -= amount;
         Address.sendValue(payable(msg.sender), amount);
     }
 
-    function depositToken(address token, uint amount) external {
+    function depositToken(address token, uint256 amount) external {
         IERC20(token).transferFrom(msg.sender, address(this), amount);
         tokenDeposits[token][msg.sender] += amount;
     }
 
-    function withdrawToken(address token, uint amount) external {
+    function withdrawToken(address token, uint256 amount) external {
         tokenDeposits[token][msg.sender] -= amount;
         IERC20(token).transfer(msg.sender, amount);
     }
